+++ src/main/kotlin/blackjack/domain/Card.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 data class Card (
     val number: Number,
